[CHORE] inject the Model specific _modelForMixin implementation into the store from the Model package
