+++ packages/store/addon/-private/system/core-store.ts
@@ -3514,6 +3514,30 @@ abstract class CoreStore extends Service {
 
     updated.length = 0;
   }
+
+  /**
+    In case someone defined a relationship to a mixin, for example:
+    ```
+      let Comment = Model.extend({
+        owner: belongsTo('commentable'. { polymorphic: true })
+      });
+      let Commentable = Ember.Mixin.create({
+        comments: hasMany('comment')
+      });
+    ```
+    we want to look up a Commentable class which has all the necessary
+    relationship metadata. Thus, we look up the mixin and create a mock
+    Model, so we can access the relationship CPs of the mixin (`comments`)
+    in this case
+
+    @method _modelForMixin
+    @protected
+    @param {String} normalizedModelName
+    @return {Record} record
+  */
+  _modelForMixin(normalizedModelName: string): unknown {
+    throw new Error('Method not implemented.');
+  }
 }
 
 defineProperty(
